refactor(examples): Split `evm` module into submodules in the `custom_node` example
