+++ examples/custom-node/src/evm/executor.rs
@@ -1,0 +1,94 @@
+use crate::evm::CustomEvmConfig;
+use alloy_evm::{
+    block::{
+        BlockExecutionError, BlockExecutionResult, BlockExecutor, BlockExecutorFactory,
+        BlockExecutorFor, ExecutableTx, OnStateHook,
+    },
+    precompiles::PrecompilesMap,
+    Database, Evm,
+};
+use alloy_op_evm::{OpBlockExecutionCtx, OpBlockExecutor, OpEvm};
+use op_revm::OpTransaction;
+use reth_ethereum::{evm::primitives::InspectorFor, node::api::ConfigureEvm};
+use reth_op::{
+    chainspec::OpChainSpec,
+    node::{OpEvmFactory, OpRethReceiptBuilder},
+    OpReceipt, OpTransactionSigned,
+};
+use revm::{
+    context::{result::ExecutionResult, TxEnv},
+    database::State,
+};
+use std::sync::Arc;
+
+pub struct CustomBlockExecutor<Evm> {
+    inner: OpBlockExecutor<Evm, OpRethReceiptBuilder, Arc<OpChainSpec>>,
+}
+
+impl<'db, DB, E> BlockExecutor for CustomBlockExecutor<E>
+where
+    DB: Database + 'db,
+    E: Evm<DB = &'db mut State<DB>, Tx = OpTransaction<TxEnv>>,
+{
+    type Transaction = OpTransactionSigned;
+    type Receipt = OpReceipt;
+    type Evm = E;
+
+    fn apply_pre_execution_changes(&mut self) -> Result<(), BlockExecutionError> {
+        self.inner.apply_pre_execution_changes()
+    }
+
+    fn execute_transaction_with_result_closure(
+        &mut self,
+        tx: impl ExecutableTx<Self>,
+        f: impl FnOnce(&ExecutionResult<<Self::Evm as Evm>::HaltReason>),
+    ) -> Result<u64, BlockExecutionError> {
+        self.inner.execute_transaction_with_result_closure(tx, f)
+    }
+
+    fn finish(self) -> Result<(Self::Evm, BlockExecutionResult<OpReceipt>), BlockExecutionError> {
+        self.inner.finish()
+    }
+
+    fn set_state_hook(&mut self, _hook: Option<Box<dyn OnStateHook>>) {
+        self.inner.set_state_hook(_hook)
+    }
+
+    fn evm_mut(&mut self) -> &mut Self::Evm {
+        self.inner.evm_mut()
+    }
+
+    fn evm(&self) -> &Self::Evm {
+        self.inner.evm()
+    }
+}
+
+impl BlockExecutorFactory for CustomEvmConfig {
+    type EvmFactory = OpEvmFactory;
+    type ExecutionCtx<'a> = OpBlockExecutionCtx;
+    type Transaction = OpTransactionSigned;
+    type Receipt = OpReceipt;
+
+    fn evm_factory(&self) -> &Self::EvmFactory {
+        self.inner.evm_factory()
+    }
+
+    fn create_executor<'a, DB, I>(
+        &'a self,
+        evm: OpEvm<&'a mut State<DB>, I, PrecompilesMap>,
+        ctx: OpBlockExecutionCtx,
+    ) -> impl BlockExecutorFor<'a, Self, DB, I>
+    where
+        DB: Database + 'a,
+        I: InspectorFor<Self, &'a mut State<DB>> + 'a,
+    {
+        CustomBlockExecutor {
+            inner: OpBlockExecutor::new(
+                evm,
+                ctx,
+                self.inner.chain_spec().clone(),
+                *self.inner.executor_factory.receipt_builder(),
+            ),
+        }
+    }
+}
