+++ examples/custom-node/src/evm/mod.rs
@@ -1,0 +1,7 @@
+mod assembler;
+mod config;
+mod executor;
+
+pub use assembler::CustomBlockAssembler;
+pub use config::CustomEvmConfig;
+pub use executor::CustomBlockExecutor;
